requirements.txt
apache-airflow-providers-amazon>=8.0.0  # For AWS integrations (S3, EMR, etc.)
boto3>=1.28.0  # AWS SDK
botocore>=1.31.0
webdriver-manager
confluent-kafka
kafka-python-ng
selenium
faker
requests==2.32.3
beautifulsoup4==4.13.3
pandas
pytz
httpcore
dotenv
gevent
discord_webhook
pyarrow>=14.0.1
pinecone>=3.0.0
sentence-transformers>=2.2.2
numpy>=1.24.3
redis>=5.0.1
psycopg2-binary
fastparquet
neo4j>=5.14.0  # For Neo4j database connections
graphdatascience